Refactor : fileUtil 로 파일 관련 함수 이동
